+++ src/client/CachedAssetsManager.ts
@@ -40,14 +40,16 @@ const contents = [
     "AvatarSkillExcelConfigData", // Skills
     "AvatarTalentExcelConfigData", // Constellations
     "AvatarPromoteExcelConfigData", // Character Ascensions
-    "AvatarCurveExcelConfigData", // Character Basic Stats
+    "AvatarCurveExcelConfigData", // Character Basic Stats Curves
     "ReliquaryExcelConfigData", // Artifacts
     "ReliquarySetExcelConfigData", // Artifact Sets
     "WeaponExcelConfigData", // Weapons
-    "EquipAffixExcelConfigData", // Artifact Set Bonus
+    "WeaponPromoteExcelConfigData", // Weapon Ascensions
+    "WeaponCurveExcelConfigData", // Weapon Basic Stats Curves
+    "EquipAffixExcelConfigData", // Artifact Set Bonus and Weapon Refinements
     "ManualTextMapConfigData", // Fight Props
     "MaterialExcelConfigData", // Materials (including NameCards)
-    "ProudSkillExcelConfigData", // Passive Talents
+    "ProudSkillExcelConfigData", // Passive Talents and Leveled Talents
     "ReliquaryAffixExcelConfigData", // Artifact Affix
     "AvatarCodexExcelConfigData", // Character Release Information
     "AvatarHeroEntityExcelConfigData", // Travelers