+++ src/models/weapon/WeaponAscension.ts
@@ -1,13 +1,13 @@
-import { JsonReader, JsonObject } from "config_file.js";
+import { JsonObject, JsonReader } from "config_file.js";
 import EnkaClient from "../../client/EnkaClient";
-import AssetsNotFoundError from "../../errors/AssetsNotFoundError";
 import UpgradeCost from "../material/UpgradeCost";
 import StatProperty, { FightProp } from "../StatProperty";
+import AssetsNotFoundError from "../../errors/AssetsNotFoundError";
 
 /**
- * @en CharacterAscension
+ * @en WeaponAscension
  */
-class CharacterAscension {
+class WeaponAscension {
     /**  */
     readonly id: number;
     /**  */
@@ -36,27 +36,27 @@ class CharacterAscension {
 
         const json = new JsonReader(this._data);
 
-        this.id = json.getAsNumber("avatarPromoteId");
+        this.id = json.getAsNumber("weaponPromoteId");
         this.ascension = json.getAsNumberWithDefault(0, "promoteLevel");
 
         this.unlockMaxLevel = json.getAsNumber("unlockMaxLevel");
 
         this.requiredAdventureRank = json.getAsNumber("requiredPlayerLevel");
 
-        this.cost = new UpgradeCost(json.getAsNumberWithDefault(0, "scoinCost"), json.has("costItems") ? json.get("costItems").mapArray((_, p) => p.getAsJsonObject()) : [], enka);
+        this.cost = new UpgradeCost(json.getAsNumberWithDefault(0, "coinCost"), json.has("costItems") ? json.get("costItems").mapArray((_, p) => p.getAsJsonObject()) : [], enka);
 
         this.addProps = json.get("addProps").filterArray((_, p) => p.has("propType") && p.has("value")).map(([, p]) => new StatProperty(p.getAsString("propType") as FightProp, p.getAsNumber("value"), enka));
     }
 
     /**
-     * @param id avatarPromoteId
+     * @param id weaponPromoteId
      * @param ascension promoteLevel
      */
-    static getById(id: number, ascension: number, enka: EnkaClient): CharacterAscension {
-        const json = enka.cachedAssetsManager.getGenshinCacheData("AvatarPromoteExcelConfigData").findArray((_, p) => p.getAsNumber("avatarPromoteId") === id && p.getAsNumberWithDefault(0, "promoteLevel") === ascension)?.[1];
-        if (!json) throw new AssetsNotFoundError("CharacterAscension", `${id}-${ascension}`);
-        return new CharacterAscension(json.getAsJsonObject(), enka);
+    static getById(id: number, ascension: number, enka: EnkaClient): WeaponAscension {
+        const json = enka.cachedAssetsManager.getGenshinCacheData("WeaponPromoteExcelConfigData").findArray((_, p) => p.getAsNumber("weaponPromoteId") === id && p.getAsNumberWithDefault(0, "promoteLevel") === ascension)?.[1];
+        if (!json) throw new AssetsNotFoundError("WeaponAscension", `${id}-${ascension}`);
+        return new WeaponAscension(json.getAsJsonObject(), enka);
     }
 }
 
-export default CharacterAscension;
+export default WeaponAscension;